Nit fixes
